Condiçoes d

+++ partials/cards_achados.php
@@ -10,9 +10,11 @@
                     <input type="hidden" class="id-card" value="<?php echo $row->id ?>">
                 </div>
                 <div class="card-action">
+                  <?php if($_COOKIE["tipo"] == 1 ): ?>
                         <?php echo '<a href="index.php?action=delete&id=' . $row->id.'">Excluir</a>' ?>
                         <?php echo '<a href="devolvidos/Fdevolvidos.php?action=devolver&id=' . $row->id.'">Devolver</a>' ?>
                         <a href="#" class="botaoAtualizar modal-atualizar">Atualizar</a>
+                  <?php endif ?>
 
                 </div>
                 <div class="card-reveal">
